feat: refactored album route and renamed image-related variables
This commit includes some major changes in the album route. Direct import path of the 'AlbumFile' from 'types' has been changed, likely to make the import clearer and to avoid confusion. The parameters in the 'routes' function was renamed to underscore to denote unused variable.

Furthermore, this commit replaces the terms related to 'image' with 'file'. This provides a more general approach, as not only images but all types of files can be considered. The endpoints for 'upload' and 'delete' have also been updated accordingly, to match with this update.